refactor count.go and didn't finish m option
